# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cash Bank apis",
        "contact": {
            "name": "Yosof Bayan",
            "url": "https://wa.me/+963967213544",
            "email": "yosofbayan75@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "local Base URL"
        },
        {
            "url": "https://cashbank-back.icrcompany.net/api/v1",
            "description": "Develop Base URL"
        }
    ],
    "paths": {
...
        "/admin/transactions": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Transaction"
                ],
                "summary": "get all transactions",
                "operationId": "717b49d75849350f2fdc8aad1e3a1b92",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "any"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "data base column name"
                        }
                    },
                    {
                        "name": "trash",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "enum": [
                                0,
                                1
                            ],
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/admin/transactions/{transaction}": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Transaction"
                ],
                "summary": "show one transaction",
                "operationId": "39bd1cd6d7dfb14dcd6465651c7ac9e7",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Transaction"
                ],
                "summary": "store new transaction",
                "operationId": "f7e08faf29cc3c01846e5e5bd6640edc",
                "parameters": [
                    {
                        "name": "transaction",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateStatusRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Users"
                ],
                "summary": "get all users",
                "operationId": "bf66fd46994d8e7fd94eb8e775fa8814",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/users/{user}": {
            "delete": {
                "tags": [
                    "Admin",
                    "Admin - Users"
                ],
                "summary": "SoftDelete an existing user",
                "operationId": "9dfab79539c83db30dbd36385b03e4eb",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
        "/admin/wallets/{id}": {
            "get": {
                "tags": [
                    "Admin",
                    "Admin - Wallet"
                ],
                "summary": "show one wallet",
                "operationId": "351882e8750316b77a525b9a0905a56b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WalletResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin",
                    "Admin - Wallet"
                ],
                "summary": "update an existing wallet",
                "operationId": "2ba1d0cc8872347492dd24849bf04f05",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_method",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "PUT",
                            "enum": [
                                "PUT"
                            ]
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateWalletRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/WalletResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": [],
                        "Accept": "json/application"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UpdateStatusRequest": {
                "required": [
                    "status"
                ],
                "properties": {
                    "status": {
                        "type": "string",
                        "enum": [
                            "cancelled",
                            "completed"
                        ]
                    },
                    "note": {
                        "type": "string"
                    },
                    "amount_confirmed": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "type": "object"
            },
            "UpdateWalletRequest": {
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "number": {
                        "type": "string"
                    },
                    "icon": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "type": "object"
            },
...
            "TransactionResource": {
                "title": "TransactionResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "WalletResource": {
                "title": "WalletResource",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}